+++ src/__tests__/App.test.js
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event'
 import { createMemoryHistory } from 'history'
 import { Router } from 'react-router-dom'
-import App from '../App';
+import App from '../components/App';
 
 describe("Testing Routes", () => {
   test.skip('renders App - Home page', () => {
