+++ src/__tests__/Home.test.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { render, screen, cleanup } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import {Home} from "../Home";
+import {Home} from "../components/Home/Home";
 
 describe("Home component", () => {
   beforeEach(() => {
