+++ src/__tests__/ListSymptoms.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, cleanup } from "@testing-library/react";
-import ListSymptoms from "../ListSymptoms";
+import ListSymptoms from "../components/Diary/ListSymptoms";
 
 const key = 1;
 const symptom = {
