+++ src/__tests__/Report.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, cleanup } from "@testing-library/react";
-import Report from "../Report";
+import Report from "../components/Report/Report";
 
 jest.mock("../ReportListView", () => {
   const ReportListView = () => <div />;
