+++ src/__tests__/ReportListItem.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, screen, cleanup } from "@testing-library/react";
-import ReportListItem from "../ReportListItem";
+import ReportListItem from "../components/Report/ReportListItem";
 
 const date = "2020-09-17";
 const symptoms = [
