+++ src/components/Diary/AdditionalData.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { toast } from "react-toastify";
-import firebase from "./firebase.js";
+import firebase from "../../firebase.js";
 
 type AdditionalDataProps = {
   userID: string | null;
