+++ src/components/Diary/ListSymptoms.tsx
