+++ src/components/Diary/MyAllergy.tsx
@@ -1,8 +1,11 @@
 import React, { useState, useEffect } from "react";
-import firebase from "./firebase.js";
-import {ReportListView} from "./ReportListView";
 
-type ReportProps = {
+import { Symptoms } from "./Symptoms";
+import { Food } from "./Food";
+import { AdditionalData } from "./AdditionalData";
+import firebase from "../../firebase.js";
+
+type MyAllergyProps = {
   userID: string | null;
 };
 
@@ -11,13 +14,18 @@ type TKids = {
   kidName: string;
 };
 
-export const Report: React.FC<ReportProps> = (props) => {
+
+
+export const MyAllergy: React.FC<MyAllergyProps> = (props) => {
   const [kids, setKids] = useState<TKids[]>([]);
 
   const [selectedKid, setSelectedKid] = useState<string>("");
   const [selectedKidId, setSelectedKidId] = useState<string>("");
 
-  const [entryMonth, setEntryMonth] =useState<string>("");
+  const [entryDate, setEntryDate] = useState<string>("");
+  const [entryMonth, setEntryMonth] = useState<string>("");
+
+  let today = new Date();
 
   const kidsRef = firebase.database().ref(props.userID + "/kids");
   const defaultKidRef = firebase.database().ref(props.userID + "/defaultKid");
@@ -60,24 +68,30 @@ export const Report: React.FC<ReportProps> = (props) => {
       setKids(newState);
     });
 
-    formatDate();
+    formatDate(today);
   }, [props]);
 
-  const formatDate = () => {
-    let mm_str;
-    let today = new Date();
+  const formatDate = (date_input: Date) => {
+    let date_output;
+    let date_output_yyyy_mm, dd_str, mm_str;
+    let dd = date_input.getDate();
 
-    let mm = today.getMonth() + 1;
-    let yyyy = today.getFullYear();
-
+    let mm = date_input.getMonth() + 1;
+    let yyyy = date_input.getFullYear();
+    
     mm_str = mm;
+    dd_str = dd;
 
+    if (dd < 10) {
+      dd_str = "0" + dd;
+    }
     if (mm < 10) {
       mm_str = "0" + mm;
     }
-    let today_month = yyyy + "-" + mm_str;
-
-    setEntryMonth(today_month);
+    date_output = yyyy + "-" +  mm_str + "-" + dd_str;
+    date_output_yyyy_mm = yyyy + "-" +  mm_str;
+    setEntryDate(date_output);
+    setEntryMonth(date_output_yyyy_mm);
   };
 
   const handleKidSelection = (e:React.ChangeEvent<any>) => {
@@ -85,21 +99,22 @@ export const Report: React.FC<ReportProps> = (props) => {
     setSelectedKidId(e.target.id);
   };
 
-  const handleEntryMonth = (e:React.ChangeEvent<any>) => {
-    setEntryMonth(e.target.value);
+  const handleEntryDate = (e:React.ChangeEvent<any>) => {
+    //sets entry date and entry month
+    setEntryDate(e.target.value);
+    setEntryMonth(e.target.value.substring(0, 7));
   };
 
   return (
     <div>
-      <div className="u-center-text u-padding-top-big  u-margin-bottom-medium">
-        <h2 data-testid="report-header" className="heading-secondary bg-color-blue ">
-          Report
-          {selectedKid != null ? (
+      <div className="u-center-text  u-padding-top-big u-margin-bottom-medium">
+        <h2 data-testid="diary-header" className="heading-secondary bg-color-blue ">
+          Diary
+          {selectedKid !== "" ? (
             <span className="u-capitalize"> - {selectedKid} </span>
           ) : null}
         </h2>
       </div>
-
       {props.userID === null ? (
         <p
           className="paragraph u-center-text u-text-color-red 
@@ -117,10 +132,11 @@ export const Report: React.FC<ReportProps> = (props) => {
         </h3>
         <div className="u-margin-bottom-small">
           {kids.map((kid, index) => (
-            <div key={kid.id}
-              className="list-kids u-capitalize"            
+            <div
+              className="list-kids u-capitalize"
+              key={kid.id}     
             >
-              <label htmlFor={kid.id}>
+              <label htmlFor={kid.id} >
                 <input
                   type="radio"
                   id={kid.id}
@@ -147,32 +163,43 @@ export const Report: React.FC<ReportProps> = (props) => {
             ) : null
           }
         </div>
-        <div className="u-text-left">
-          <h3
-            className="heading-tertiary 
-          u-margin-bottom-small"
+        <div className="u-text-left u-margin-bottom-small">
+          <label
+            htmlFor="entryDate"
+            className="heading-tertiary u-margin-right"
           >
-            Choose Month
-          </h3>
-
+            Date
+          </label>
           <input
-            type="month"
-            name="entryMonth"
-            value={entryMonth}
-            onChange={(e) => handleEntryMonth(e)}
+            type="date"
+            name="entryDate"
+            value={entryDate}
+            onChange={(e) => handleEntryDate(e)}
             required
-            data-testid="month-selector"
           />
         </div>
       </div>
 
-      {selectedKid != null ? (
-        <ReportListView
-          userID={props.userID}
-          kidId={selectedKidId}
-          date_yyyy_mm={entryMonth}
-        />
-      ) : null}
+      <Symptoms
+        userID={props.userID}
+        kidId={selectedKidId}
+        date={entryDate}
+        date_yyyy_mm={entryMonth}
+      />
+
+      <Food
+        userID={props.userID}
+        kidId={selectedKidId}
+        date={entryDate}
+        date_yyyy_mm={entryMonth}
+      />
+
+      <AdditionalData
+        userID={props.userID}
+        kidId={selectedKidId}
+        date={entryDate}
+        date_yyyy_mm={entryMonth}
+      />
     </div>
   );
 };
