+++ src/components/Report/ReportListSymptoms.tsx
